+++ src/Resources/lang/de/app.php
@@ -1,0 +1,19 @@
+<?php
+
+return [
+    'system' => [
+        'title'                   => 'MultiSafePay',
+        'info'                    => 'Bieten Sie die beliebtesten Zahlungsmethoden an, um Ihre Verkäufe lokal und international zu steigern',
+        'api-key'                 => 'API-Schlüssel',
+        'production'              => 'Produktionsmodus',
+        'order-id-prefix'         => 'Bestell-ID-Präfix',
+        'display_cart_items'      => 'Warenkorb anzeigen',
+        'display_cart_items_info' => 'Zeigen Sie Artikel aus dem Warenkorb des Kunden auf der Zahlungsseite an',
+    ],
+    'shop' => [
+        'order_already_paid'       => 'Diese Bestellung ist bereits bezahlt.',
+        'pay-button'               => 'Jetzt bezahlen',
+        'payment_success'          => 'Vielen Dank für Ihre Zahlung in Höhe von :amount.',
+        'payment_processing_error' => 'Fehler bei der Verarbeitung der Zahlung.',
+    ],
+];